Spec refactor - split provider spec in two
File was too big.
